requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
pyserial>=3.5
numpy>=1.24.0
sounddevice>=0.4.6
keyboard>=0.13.5
pyttsx3>=2.90
ollama>=0.1.0
